refactor(ipc): split lib/ipc.ts into domain modules behind a barrel — behavior unchanged
- Delete src/lib/ipc.ts (1765 lines)
- Create src/lib/ipc/ with 22 domain modules + barrel index.ts
- Barrel re-exports every symbol the old ipc.ts exported (same names, same signatures)
- All existing `import { ... } from '@/lib/ipc'` paths continue to resolve identically
- No logic changes, no signature changes, no renames
- Mock state mutation (mockSettings, mockStylePacks) centralised in mock-data.ts
  via exported helpers (mockSetSettings, mockSetDefaultPolishMode, etc.) to avoid
  ES module read-only binding violations
- TypeScript: zero errors (npx tsc --noEmit)
- Build: clean (npm run build)
