+++ openless-all/app/src/lib/ipc/platform-exports.ts
@@ -1,0 +1,9 @@
+export { isAndroid, isDesktop, isMobile } from "../platform"
+export {
+    getAndroidOverlayStatus,
+    requestAndroidOverlayPermission,
+    showAndroidOverlay,
+    hideAndroidOverlay,
+    getAndroidAccessibilityStatus,
+    requestAndroidAccessibilityPermission,
+} from '../../../android/frontend/lib/androidIpc'
